Create AccessibillityService and AccessibillityType for enum and AccessibillityController

+++ backend/src/main/java/com/main/app/repository/AccessibillityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccessibillityRepository extends JpaRepository<Accessibillity, Integer> {
+public interface AccessibillityRepository extends JpaRepository<Accessibillity, Long> {
 }
